Update README.md
RAWINVESTING


DESCRIPTIONS:
RawInvesting is a primary sector investment stock market website focused on Cotton,coffee, corn, sugar and wheat. With a minimalist and readable interface in a simple and very clear user-oriented way.

This informative website is visually attractive (look), and the user will feel comfortable when interacting with the platform (feel) because each attribute that has been created and designed on the platform has a purpose (usability) and several Call To Action to make the purchase and sale of commodities.

Rawinvesting was born with the idea of responding to the need to delve into a very specific market niche that today is not found on the Internet, with the aim of positioning itself in the very near future as the reference website for information in the financial and investment field for Europe and the United States, but with the goal of becoming the first for Spain and Latin America by 2025.

For this reason the website is in an initial process (Beta) that will be updated and improved as time goes by and as market needs evolve.

As we have highlighted in our buyer persona, the profile of our future users of the website are Technical Stock Market Investors in the natural resources sector:
Investment Banks
Private investors
Foreign trade technicians
Small investors.


2)WEB MAP

Our website is divided into 5 views:


HOME:
This view is the general view of the website and shows the last month’s price of each of the commodities that make up our website and shows and you can also see a graph showing how their prices have a evolved over  the last year.
It includes a section for reviews from investors or users of the site to be able to comment or include any opinion about the site.
-DETAILS
This view shows a unit chart for each of the commodities.You can select how you want to view the chart, either as a line or bar chart, and it can also include the months of a particular year, or a range of years to get the evolution of prices.In addition, it includes a cheese chart that shows the percentage of investment in the last year, made by investors, in each of the commodities.
-CUSTOMIZE
 This view gives the user the possibility to select to compare the commodities he/she wants and will show in a graph the prices given a period of time, so that the historical evolution can be seen in the graph.In addition, at the bottom, it includes a part of frequently asked questions for the user
-BLOG
This view collects news about investments of each of the agricultural commodities that are on the page
-ABOUT US
This view is a contact page so that the user can get in touch with the website and it also shows the people who have created the website
- COOKIES POLICY
Section dedicated to cookies policy
-PRIVACY POLICY
Section dedicated to privacy policy
-LEGAL NOTICE
Section dedicated to Legal Notice

3)NAVBAR

This section is included in all pages of the website. It includes responsive design.
This navigation bar is included in the side part of the browser.
You also have access to the different parts of the website: Home,Details,Customize,blog, About Us,Cookies policy, Privacy Policy and Legal Notice.





4)RAWINVESTING’S FOUNDERS  


Nasreddin Oufallah
FullStack Development


Linkedin: https://www.linkedin.com/in/nasreddin-oufallah/
GitHub: https://github.com/nasre21




Zucelys Guedes
FullStack Development


Linkedin: https://www.linkedin.com/in/zguedest/
GitHub: https://github.com/ZGuedest








Lucas Ibar
FullStack Development


Linkedin: https://www.linkedin.com/in/lucas-ibar-basaldua/
GitHub: https://github.com/Lucasibf5






Jessica Mosteiro
FullStack Development


Linkedin: https://www.linkedin.com/in/jessica-mosteiro/
GitHub: https://github.com/JJessa






Maria Diez
FullStack Development


Linkedin: https://www.linkedin.com/in/mariadiezesteve/
GitHub: https://github.com/MariaDiezEsteve






Gorka Ibarrondo
FullStack Development


Linkedin: https://www.linkedin.com/in/gorka-ibarrondo-busturia-21a317259/
GitHub: https://github.com/Gorka1999